add cuda check
